+++ src/serializing.rs
@@ -1,1 +1,4 @@
 mod md;
+
+#[cfg(feature = "markdown")]
+pub(crate) use md::serialize_md;